+++ src/game.rs
@@ -5,5 +5,5 @@ pub trait Game: Hash {
     type Children;
 
     fn branch(&self) -> Vec<Self::Children>;
-    fn state(&self) -> Array1<f64>;
+    fn state(&self) -> Array1<f32>;
 }
